move the PrefsListener out to its own class
