+++ topl-service/src/main/java/io/matthewnelson/topl_service/service/TorService.kt
@@ -16,9 +16,8 @@ import io.matthewnelson.topl_service.onionproxy.ServiceEventBroadcaster
 import io.matthewnelson.topl_service.onionproxy.ServiceEventListener
 import io.matthewnelson.topl_service.onionproxy.ServiceTorInstaller
 import io.matthewnelson.topl_service.onionproxy.ServiceTorSettings
-import io.matthewnelson.topl_service.util.ServiceConsts.PrefKeyBoolean
+import io.matthewnelson.topl_service.prefs.TorServicePrefsListener
 import io.matthewnelson.topl_service.util.ServiceConsts.ServiceAction
-import io.matthewnelson.topl_service.util.TorServicePrefs
 import kotlinx.coroutines.*
 
 internal class TorService: Service() {
@@ -58,6 +57,7 @@ internal class TorService: Service() {
     }
 
     private lateinit var broadcastLogger: BroadcastLogger
+    private lateinit var torServicePrefsListener: TorServicePrefsListener
 
     override fun onBind(intent: Intent?): IBinder? {
         return null
@@ -66,14 +66,13 @@ internal class TorService: Service() {
     override fun onCreate() {
         super.onCreate()
         ServiceNotification.get().startForegroundNotification(this)
-        initTorServicePrefsListener(this)
         initTOPLCore(this)
         broadcastLogger = onionProxyManager.createBroadcastLogger(TorService::class.java)
-        torServicePrefsListener.initBroadcastLogger(onionProxyManager)
+        broadcastLogger.notice("BuildConfig.DEBUG set to: $buildConfigDebug")
+        torServicePrefsListener = TorServicePrefsListener(this, onionProxyManager)
     }
 
     override fun onDestroy() {
-        broadcastLogger.debug("onDestroy called. Cleaning up.")
         torServicePrefsListener.unregister()
         supervisorJob.cancel()
         super.onDestroy()
@@ -86,7 +85,7 @@ internal class TorService: Service() {
 
     override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
         intent?.action?.let {
-            broadcastLogger.debug("Executing ServiceAction: $it")
+            broadcastLogger.debug("Received ServiceAction: $it")
             executeAction(it)
         }
         return START_STICKY
@@ -204,75 +203,5 @@ internal class TorService: Service() {
                 }
             }
         }
-    }
-
-
-    ///////////////////////
-    /// TorServicePrefs ///
-    ///////////////////////
-    private lateinit var torServicePrefsListener: TorServicePrefsListener
-
-    private fun initTorServicePrefsListener(torService: TorService) {
-        torServicePrefsListener = TorServicePrefsListener(TorServicePrefs(torService))
-    }
-
-    /**
-     * Gets initialized in [onCreate] **before** [initTOPLCore] is called so that the
-     * initial [onSharedPreferenceChanged] proc won't do anything.
-     *
-     * Listens to [TorServicePrefs] for changes such that while Tor is running, it can
-     * query [onionProxyManager] to have it updated immediately (if the setting doesn't
-     * require a restart).
-     *
-     * @param [torServicePrefs] Our target to listen to for any changes
-     * */
-    private inner class TorServicePrefsListener(
-        val torServicePrefs: TorServicePrefs
-    ): SharedPreferences.OnSharedPreferenceChangeListener {
-
-        private var broadcastLogger: BroadcastLogger? = null
-        fun initBroadcastLogger(onionProxyManager: OnionProxyManager) {
-            if (broadcastLogger == null)
-                broadcastLogger =
-                    onionProxyManager.createBroadcastLogger(TorServicePrefsListener::class.java)
-        }
-
-        // Register itself immediately upon instantiation.
-        init {
-            torServicePrefs.registerListener(this)
-        }
-
-        /**
-         * Called from [onDestroy] to prevent memory leaks.
-         * */
-        fun unregister() {
-            broadcastLogger?.debug("Unregistering self")
-            torServicePrefs.unregisterListener(this)
-        }
-
-        override fun onSharedPreferenceChanged(sharedPreferences: SharedPreferences?, key: String?) {
-            if (::onionProxyManager.isInitialized && !key.isNullOrEmpty()) {
-                broadcastLogger?.debug("$key was modified")
-                when (key) {
-                    PrefKeyBoolean.HAS_DEBUG_LOGS -> {
-                        // TODO: Think about doing something with local prefs such that it
-                        //  turns Debugging off at every application start automatically?
-                        //  .
-                        //  Especially necessary if I switch Tor's debug output location from
-                        //  SystemOut to log to a file (more secure).
-                        //  .
-                        //  Will need to create another class available to Library user
-                        //  strictly for Tor logs if logging to a file, such that they can
-                        //  easily query, read, and load them to views.
-                        //  Maybe a `TorDebugLogHelper` class?
-                        //  .
-                        //  Will need some way of automatically clearing old log files, too.
-                        if (!onionProxyManager.torStateMachine.isOff)
-                            onionProxyManager.refreshBroadcastLoggersHasDebugLogsVar()
-                    }
-                }
-            }
-        }
-
     }
 }
